+++ packages/backend/src/engine/backend-secret/backend_secret.ts
@@ -1,4 +1,8 @@
-import { BackendIdentifier, BackendSecret } from '@aws-amplify/plugin-types';
+import {
+  BackendIdentifier,
+  BackendSecret,
+  ResolvePathResult,
+} from '@aws-amplify/plugin-types';
 import { Construct } from 'constructs';
 import { BackendSecretFetcherFactory } from './backend_secret_fetcher_factory.js';
 import { SecretValue } from 'aws-cdk-lib';
@@ -35,10 +39,16 @@ export class CfnTokenBackendSecret implements BackendSecret {
   /**
    * Resolve to the secret path
    */
-  resolvePath = (backendIdentifier: BackendIdentifier): string => {
-    return ParameterPathConversions.toParameterFullPath(
-      backendIdentifier,
-      this.name
-    );
+  resolvePath = (backendIdentifier: BackendIdentifier): ResolvePathResult => {
+    return {
+      branchSecretPath: ParameterPathConversions.toParameterFullPath(
+        backendIdentifier,
+        this.name
+      ),
+      sharedSecretPath: ParameterPathConversions.toParameterFullPath(
+        backendIdentifier.namespace,
+        this.name
+      ),
+    };
   };
 }